chore: adjust importing file text-editor
