+++ src/components/common/TextEditor/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./TextEditor";
+export { default as Toolbar } from "./Toolbar";
+export { default as ToolbarButton } from "./ToolbarButton";
